+++ src/main/java/com/example/restaurantapi/dto/item/CreateItemDto.java
@@ -10,9 +10,9 @@ public class CreateItemDto {
     public Double quantity;
     public Double price;
     public String unit;
-    public Long restaurantId;
+    public int restaurantId;
     public String restaurantName;
-    public Long itemType;
+    public int itemType;
     public String itemTypeName;
 
     public static CreateItemDto of (Item item) {
