+++ src/main/java/com/example/restaurantapi/repository/ConfirmationTokenRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Integer> {
     Optional<ConfirmationToken> findConfirmationTokenByConfirmationToken(String token);
 }
