+++ src/main/java/com/example/restaurantapi/repository/ItemRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
     List<Item> findByTitleContaining(String name);
-    List<Item> findByItemTypeId(Long id);
+    List<Item> findByItemTypeId(int id);
     List<Item> findByRestaurant(Restaurant restaurant);
 
 }
