+++ src/main/java/com/example/restaurantapi/repository/ItemTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.restaurantapi.repository;
 import com.example.restaurantapi.model.ItemType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemTypeRepository extends JpaRepository<ItemType, Long> {
+public interface ItemTypeRepository extends JpaRepository<ItemType, Integer> {
 }
