+++ src/main/java/com/example/restaurantapi/repository/MenuTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MenuTypeRepository extends JpaRepository<MenuType, Long> {
+public interface MenuTypeRepository extends JpaRepository<MenuType, Integer> {
 
 }
