+++ src/main/java/com/example/restaurantapi/repository/OpeningPeriodRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface OpeningPeriodRepository extends JpaRepository<OpeningPeriod, Long> {
+public interface OpeningPeriodRepository extends JpaRepository<OpeningPeriod, Integer> {
     List<OpeningPeriod> findByRestaurant(Restaurant restaurant);
 }
