+++ src/main/java/com/example/restaurantapi/repository/RestaurantTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.restaurantapi.repository;
 import com.example.restaurantapi.model.RestaurantType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantTypeRepository extends JpaRepository<RestaurantType, Long> {
+public interface RestaurantTypeRepository extends JpaRepository<RestaurantType, Integer> {
 }
